DEV: Use RAILS_ENV=test for theme qunit tests (#121)
There are subtle differences in the test/development environments when creating/migrating the database. Using the development environment here would sometimes lead to behavior which could not be reproduced in the plugin's own CI.

Instead, we can use the test environment, and allow the `theme:qunit` task to switch to the development environment when booting the server. `DISCOURSE_DEV_DB` allows us to use the prepared test database for that server.