DEV: Add ember 5 to test matrix (#126)
Discourse is upgrading to Ember 5. This commit will add Ember 5 jobs to the theme/plugin test matrices.

For now, these are 'optional'. Failures in the tests will not cause the whole CI suite to fail. Unfortunately GitHub doesn't provide native support for that kind of thing in workflows. So instead, we're setting tests to continue-on-failure: true and then using the GitHub 'Checks' API to report a separate 'neutral' status check if they fail.

That neutral status check includes links to the original logs, and the upgrade docs on Meta.