DEV: Remove Chrome before running tests
Why this change?

This is a follow-up to https://github.com/discourse/discourse/commit/99921120a1fe2d59e70036c85486c81325f36715
where both Chrome and Chromium are now installed in the test image as we
transition to using Chromium only in our test which supports both x86
and arm64 with binaries.