DEV: Add MinIO step for system specs (#159)
System specs that depend on S3 uploads are currently not supported on CI for plugins because MinIO, which is required for S3 system specs, isn't installed in the plugin tests workflow. This PR adds steps to install MinIO to support S3 system specs. See https://github.com/discourse-org/discourse-new-features-feeds/pull/47 for an example.